moving task types from api.type
